[DOC][VPR]Add new integration plugin header file to the Versioned Plugin Reference (#891)
Integration plugins need a header that is different from the stand-alone plugin header.
The plugin docs should to point to the integration repo rather than the input,
output, filter, or codec repo. The new header also adds boilerplate text to indicate that
the individual plugin is part of an integration rather than stand-alone. This work implements
those changes.