Doc:Fix links to es broken at 7.9 cutover (#893)
Elasticsearch doc improvements and restructuring broke some Logstash links at 7.9 and before.
This problem is recurring because we use /current/ in links as a work around for versioning incompatibilities between the stack and individual plugins. When current resolves to a new branch, any link targets that no longer exist are broken. For example, if a link exists in 7.7 but not in 7.8, it will break when /current/=7.8. 

In this PR I added links to more appropriate versions. These files are generated, and so fixing the output should make for a more evergreen fix over just changing the link to resolve for current.